+++ src/parser/nodes.py
@@ -491,15 +491,18 @@ class ExportNode(Node):
     """Node for `export` structure.
     identifier is the name of the module to import. It is a token. Example: Token(TT_IDENTIFIER, 'lorem_ipsum')
     """
-    def __init__(self, expr: Node, as_identifier: Token, pos_start, pos_end):
-        self.expr: Node = expr
-        self.as_identifier: Token = as_identifier
+    def __init__(self, expr_or_identifier: Node | Token, as_identifier: Token | None, pos_start, pos_end):
+        self.expr_or_identifier: Node | Token = expr_or_identifier
+        self.as_identifier: Token | None = as_identifier
 
         self.pos_start = pos_start
         self.pos_end = pos_end
 
     def __repr__(self):
-        return f"export:{self.expr}:as:{self.as_identifier}"
+        if self.as_identifier is None:
+            return f"export:{self.expr_or_identifier}"
+        else:
+            return f"export:{self.expr_or_identifier}:as:{self.as_identifier}"
 
 
 # FILE NODES